moved common functions to utils
