+++ packages/web-console/src/utils/index.ts
@@ -31,3 +31,5 @@ export * from "./styled"
 export * from "./formatSql"
 export * from "./platform"
 export * from "./pick"
+export * from "./fetchUserLocale"
+export * from "./getLocaleFromLanguage"
